+++ tests/functional/translate/cram/_setup.sh
@@ -1,2 +1,0 @@
-pushd "$TESTDIR/../../" > /dev/null
-export AUGUR="${AUGUR:-../../bin/augur}"