+++ tests/functional/translate/cram/translate-with-genbank.t
@@ -1,18 +1,21 @@
 Setup
 
-  $ pushd "$TESTDIR" > /dev/null
-  $ source _setup.sh
+  $ export AUGUR="${AUGUR:-$TESTDIR/../../../../bin/augur}"
+  $ export DATA="$TESTDIR/../data"
+  $ export SCRIPTS="$TESTDIR/../../../../scripts"
 
 Translate amino acids for genes using a GenBank file.
 
   $ ${AUGUR} translate \
-  >   --tree translate/data/zika/tree.nwk \
-  >   --ancestral-sequences translate/data/zika/nt_muts.json \
-  >   --reference-sequence translate/data/zika/zika_outgroup.gb \
+  >   --tree "$DATA/zika/tree.nwk" \
+  >   --ancestral-sequences "$DATA/zika/nt_muts.json" \
+  >   --reference-sequence "$DATA/zika/zika_outgroup.gb" \
   >   --genes CA PRO \
-  >   --output-node-data $TMP/aa_muts.json
-  Validating schema of 'translate/data/zika/nt_muts.json'...
+  >   --output-node-data aa_muts.json
   Read in 3 features from reference sequence file
+  Validating schema of '.+nt_muts.json'... (re)
   amino acid mutations written to .* (re)
-  $ python3 "../../scripts/diff_jsons.py" translate/data/zika/aa_muts_genbank.json $TMP/aa_muts.json
+
+  $ python3 "$SCRIPTS/diff_jsons.py" $DATA/zika/aa_muts_genbank.json aa_muts.json \
+  >  --exclude-regex-paths "root\['annotations'\]\['.+'\]\['seqid'\]"
   {}